move chip out of the form folder and adjust references
