+++ src/core/Chip/Chip.md
